AR-733: Refactored code to avoid duplication
